BACKEND/requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.3
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.4
django-cors-headers==4.2.0
django-import-export==3.3.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.0
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0